requirements.txt
langchain==1.0.3
langchain-core==1.0.2
langchain-community==0.4.1
langchain-groq==1.0.0
feedparser
python-dotenv
